third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbdatabase_transaction5-expected.txt

This is a testharness.js-based test.
PASS IDBDatabase.transaction() - If storeNames is an empty list, the implementation must throw a DOMException of type InvalidAccessError 
Harness: the test ran to completion.
